requirements_eval.txt
# Requirements for LLM GAN Evaluation System

# Core dependencies (should already be installed)
numpy>=1.21.0
pandas>=1.3.0

# OpenAI API
openai>=1.0.0

# Statistical analysis and visualization
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: For loading .env files
python-dotenv>=0.19.0